refactor(content): folder-ize logic modules (Phase 3)
Phase 3 of the WKC migration. Moves single-file modules under src/logic/
into per-component folders following the standard component.ts +
optional types.ts + index.ts split.

Folder-ized:

- logic/hashing.ts -> logic/hashing/
- logic/entity-parser.ts -> logic/entity-parser/
- logic/query-params.ts -> logic/query-params/
- logic/erc721.ts -> logic/erc721/
- logic/snapshots.ts -> logic/snapshots/
- logic/deployments.ts -> logic/deployments/ (IDeploymentsComponent moved
  to types.ts)
- logic/checker.ts -> logic/checker/
- logic/third-party-contract-registry.ts -> logic/third-party-contract-registry/
  (ContractType enum + ThirdPartyContractRegistry type moved to types.ts)
- logic/third-party-item-checker.ts -> logic/third-party-item-checker/

Each new folder has an index.ts that re-exports the public surface, so
all 49 existing import sites resolve unchanged via Node's directory
module resolution. No importer paths needed updating.

Internal relative imports inside each moved file adjusted by one path
level (e.g. ../types -> ../../types).

Stays flat (small helpers, lifecycle/infra, or reshape coming in later
phases): formatters.ts, time-range.ts, contract-helpers.ts,
config-helpers.ts, status-checks.ts, entities.ts, synchronization.ts,
delete-unreferenced-files.ts, components-lifecycle.ts.

tsc clean, unit suite green (213 tests, no diff).
