+++ content/src/logic/deployments/index.ts
@@ -1,0 +1,16 @@
+export {
+  isEntityDeployed,
+  retryFailedDeploymentExecution,
+  mapDeploymentsToEntities,
+  saveDeploymentAndContentFiles,
+  calculateOverwrites,
+  MAX_HISTORY_LIMIT,
+  getCuratedOffset,
+  getCuratedLimit,
+  buildDeploymentFromHistoricalDeployment,
+  buildHistoricalDeploymentsFromRow,
+  getDeployments,
+  getDeploymentsForActiveEntities,
+  createDeploymentsComponent
+} from './component'
+export type { IDeploymentsComponent } from './types'
