+++ content/src/logic/deployments/types.ts
@@ -1,0 +1,6 @@
+import { Deployment } from '../../deployment-types'
+
+export interface IDeploymentsComponent {
+  getDeploymentsForActiveThirdPartyItemsByEntityIds(entityIds: string[]): Promise<Deployment[]>
+  updateMaterializedViews(): Promise<void>
+}
