+++ content/src/logic/entity-parser/index.ts
@@ -1,0 +1,1 @@
+export { getEntityFromBuffer } from './component'
