+++ content/src/logic/erc721/index.ts
@@ -1,0 +1,1 @@
+export { buildUrn, formatERC21Entity } from './component'
