+++ content/src/logic/hashing/component.ts
