+++ content/src/logic/hashing/index.ts
@@ -1,0 +1,1 @@
+export { calculateIPFSHashes, calculateDeprecatedHashes } from './component'
