+++ content/src/logic/snapshots/index.ts
@@ -1,0 +1,1 @@
+export { generateAndStoreSnapshot, generateSnapshotsInMultipleTimeRanges } from './component'
