+++ content/src/logic/third-party-contract-registry/index.ts
@@ -1,0 +1,3 @@
+export { createThirdPartyContractRegistry, loadCacheFile, saveCacheFile } from './component'
+export { ContractType } from './types'
+export type { ThirdPartyContractRegistry } from './types'
