+++ content/src/logic/third-party-item-checker/index.ts
@@ -1,0 +1,1 @@
+export { createThirdPartyItemChecker } from './component'
